# NewCompressReader: Separate buffers to LZ4_compress_fast_continue