Ensure we do not trigger the bug where the input buffers to
LZ4_compress_fast_continue are "accidentally" contiguous, so it
compresses blocks across them. This bug was fixed for Writer in
commit ca071684a5c2.

I do not believe CompressReader can trigger this bug because it will
only call LZ4_compress_fast_continue with a "partial" block < 64 kiB
once, at the end of the file. To trigger the problem, we need two
"partial" blocks, because then LZ4_compress_fast_continue will assume
it can use the last 64 kiB of data, which spans more than just the
last call. However, it seems worth explicitly separating these
buffers to make this impossible, in case the code is changed in the
future.

The benchmark results seem to be slightly faster after this change
on my laptop. If true, this must be from malloc/free/Cgo overhead.

Before (best run of 5 with -benchtime=30s):

```
goos: darwin
goarch: amd64
pkg: github.com/DataDog/golz4
cpu: Intel(R) Core(TM) i7-8569U CPU @ 2.80GHz
BenchmarkStreamCompressReader-8        43509      795162 ns/op  13186.95 MB/s      248 B/op        5 allocs/op
```

After (best run of 5)

```
BenchmarkStreamCompressReader-8        45981      772836 ns/op  13567.90 MB/s      248 B/op        5 allocs/op
```